fear : 주문 생성 및 삭제

+++ src/main/java/storePractice/store/domain/Order.java
@@ -11,7 +11,7 @@ public class Order {
 
     @Id
     @GeneratedValue
-    private int Id;
+    private Long Id;
 
     @OneToMany
     private List<OrderItem> orderItems;
